refactor(flows): consolidate common/ files - eliminate 4 internal-only modules
Merge single-caller abstractions into their consumers:
- buildRunFlow.ts → createAgentRunFlow.ts (inline flow wiring)
- AgentInitNode.ts → createAgentRunFlow.ts (inline init node class)
- finalizeLifecycle.ts → createFinalizeNode.ts (inline error handling)
- nodeExecution.ts → types.ts + createFinalizeNode.ts (move types)

Remove re-exports - all imports now from source:
- AgentRunHooks imported directly from @agent/core/IAgent
- No more indirection through common/types.ts

Result: 10 files → 6 files, cleaner dependency graph
