+++ src/agent/implementations/flows/common/index.ts
@@ -1,9 +1,18 @@
-export * from './AgentInitNode';
+/**
+ * Common agent flow infrastructure.
+ *
+ * Exports:
+ * - AgentLifecycle: Phase/status state machine
+ * - AgentRunFlowRunner: Flow execution with hooks
+ * - createAgentRunFlow: Flow factory (includes init node, link wiring)
+ * - createStandardFinalizeNode: Standard finalize node factory
+ * - types: AgentRunShared, BaseFlowShared
+ * - runStateSchemas: Serialization schemas
+ */
+
 export * from './AgentLifecycle';
 export * from './AgentRunFlowRunner';
-export * from './buildRunFlow';
 export * from './createAgentRunFlow';
 export * from './createFinalizeNode';
-export * from './nodeExecution';
 export * from './runStateSchemas';
 export * from './types';
